+++ backend/src/services/dashboard/overview/topVisited.ts
@@ -1,0 +1,11 @@
+import { findTopVisited } from "@/models/dashboard/overview";
+
+export const retrieveTopVisited = async (limit: number) => {
+  const retrievedTopVisited = await findTopVisited(limit);
+  return retrievedTopVisited.map(
+    ({ pageLocation, _count: { pageLocation: count } }) => ({
+      pageLocation,
+      count,
+    }),
+  );
+};
